Spread out HW functions to sep files
